requirements.txt
amqp==2.5.2
billiard==3.6.3.0
boto==2.49.0
boto3==1.12.21
botocore==1.15.21
celery==4.4.2
certifi==2019.11.28
chardet==3.0.4
defusedxml==0.6.0
dj-database-url==0.5.0
Django==2.2.7
django-allauth==0.40.0
django-celery-beat==2.0.0
django-celery-results==1.2.1
django-cors-headers==3.2.1
django-heroku==0.3.1
django-rest-auth==0.9.5
django-timezone-field==4.0
djangorestframework==3.10.3
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.4.0
docutils==0.15.2
fuzzywuzzy==0.18.0
gensim==3.8.1
gunicorn==20.0.4
ibm-cloud-sdk-core==1.5.1
ibm-watson==4.5.0
idna==2.9
importlib-metadata==1.6.0
jmespath==0.9.5
kombu==4.6.8
numpy==1.18.1
oauthlib==3.1.0
pandas==1.0.1
psycopg2==2.8.4
PyJWT==1.7.1
python-crontab==2.4.2
python-dateutil==2.8.1
python-http-client==3.3.0
python-Levenshtein==0.12.0
python-memcached==1.59
python3-openid==3.1.0
pytz==2019.3
redis==3.5.0
requests==2.23.0
requests-oauthlib==1.3.0
s3transfer==0.3.3
scipy==1.4.1
sendgrid==6.4.6
six==1.14.0
smart-open==1.9.0
sqlparse==0.3.1
starkbank-ecdsa==1.0.0
urllib3==1.25.8
vine==1.3.0
websocket-client==0.48.0
whitenoise==5.0.1
zipp==3.1.0
